feat: finalização do metodo delete user

+++ src/useCases/User/DeleteUserUseCase/DeleteUserUseCase.ts
@@ -1,6 +1,7 @@
 import { IUserRepository } from "@/repositories/User/IUserRepository";
 import { DeleteUserDTO } from "./DeleteUserDTO";
 import { ErrorUserAlreadyNotExist } from "@/erros/ErrorUserAlreadyExist";
+import { ErrorWithoutPermission } from "@/erros/ErrorWithoutPermission";
 
 export class DeleteUserUseCase {
   constructor(private userRepository: IUserRepository) {}
@@ -10,6 +11,10 @@ export class DeleteUserUseCase {
 
     if (!userAlreadyExist) {
       throw new ErrorUserAlreadyNotExist();
+    }
+
+    if(userAlreadyExist.isOwner == 'Owner'){
+      throw new ErrorWithoutPermission()
     }
 
     await this.userRepository.delete(id);
